# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ubora API Documentation",
        "description": "Documentation de l'API Ubora Backend",
        "contact": {
            "name": "Support Ubora",
            "email": "support@ubora.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://ubora.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API Ubora"
        }
    ],
    "paths": {
...
        "/api/periods": {
            "get": {
                "tags": [
                    "Périodes"
                ],
                "summary": "Lister les périodes académiques",
                "description": "Récupère la liste paginée des périodes académiques avec possibilité de recherche et filtrage.",
                "operationId": "listPeriods",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par année",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2024"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "dispatch"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5,
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des périodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PeriodResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Périodes"
                ],
                "summary": "Créer une nouvelle période académique",
                "description": "Crée une nouvelle période académique. Si aucune année n'est spécifiée, utilise l'année courante.",
                "operationId": "createPeriod",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "year": {
                                        "description": "Année de début de la période (optionnel)",
                                        "type": "integer",
                                        "example": 2025
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Période créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PeriodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création de la période."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getYearsPeriod": {
            "get": {
                "tags": [
                    "Périodes"
                ],
                "summary": "Obtenir la liste des années avec périodes",
                "description": "Récupère toutes les périodes au format 'année1-année2' pour les sélecteurs/dropdowns.",
                "operationId": "getYearsPeriod",
                "responses": {
                    "200": {
                        "description": "Liste des années avec périodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "year": {
                                                "type": "string",
                                                "example": "2024-2025"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/periods/{id}": {
            "get": {
                "tags": [
                    "Périodes"
                ],
                "summary": "Obtenir les détails d'une période",
                "description": "Récupère les informations détaillées d'une période académique spécifique.",
                "operationId": "getPeriod",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la période",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la période",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PeriodResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Période non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Période non trouvée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/periods/{period}/status": {
            "put": {
                "tags": [
                    "Périodes"
                ],
                "summary": "Changer le statut d'une période",
                "description": "Met à jour le statut d'une période académique selon le workflow prédéfini (dispatch → présélection → sélection → close).",
                "operationId": "changePeriodStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la période",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Nouveau statut",
                                        "type": "string",
                                        "enum": [
                                            "dispatch",
                                            "preselection",
                                            "selection",
                                            "close"
                                        ],
                                        "example": "preselection"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Statut mis à jour avec succès."
                                        },
                                        "period": {
                                            "$ref": "#/components/schemas/PeriodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Modification non autorisée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Modification non autorisée pour ce statut."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Période non trouvée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PeriodResource": {
                "title": "Period Resource",
                "description": "Ressource de période académique formatée",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "year": {
                        "type": "string",
                        "example": "2024-2025"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "dispatch",
                            "preselection",
                            "selection",
                            "close"
                        ],
                        "example": "dispatch"
                    },
                    "evaluators_count": {
                        "type": "integer",
                        "example": 15
                    },
                    "criteria_count": {
                        "type": "integer",
                        "example": 8
                    },
                    "candidats_count": {
                        "type": "integer",
                        "example": 250
                    },
                    "preselection_count": {
                        "type": "integer",
                        "example": 180
                    },
                    "selection_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-11T10:42:29.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-11T10:42:29.000000Z"
                    },
                    "progression_preselection": {
                        "type": "string",
                        "example": "72%"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token JWT obtenu lors de l'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}